// types.ts
export enum InsuranceProductStatus {
    Active = "Active",
    Pilot = "Pilot",
    InReview = "In Review",
    Discontinued = "Discontinued"
}

export type InsuranceCategory = 'Motor' | 'Medical' | 'Marine' | 'Pet';

export interface InsuranceProduct {
    id: string;
    name: string;
    iconUrl: string;
    status: InsuranceProductStatus;
    underwriter: {
        name: string;
        email: string;
    };
    lastUpdate: string; // ISO date string
    category: InsuranceCategory;
    policyCode: string;
    description: string;
    keyFeatures: string[];
}

// Type for creating a new product. Omits fields generated by the system/DB.
export type NewInsuranceProduct = Omit<InsuranceProduct, 'id' | 'underwriter' | 'lastUpdate'>;